//Mahesh

+++ src/main/java/com/ats/adminpanel/model/GetSellBillDetail.java
@@ -15,13 +15,11 @@ public class GetSellBillDetail {
 
 	private float grandTotal;
 	
-
 	private float mrp;
 	
 	private float mrpBaseRate;
-	
 
-	private int qty;
+	private float qty;
 	
 	private float sgstPer;
 	
@@ -77,11 +75,11 @@ public class GetSellBillDetail {
 		this.mrp = mrp;
 	}
 
-	public int getQty() {
+	public float getQty() {
 		return qty;
 	}
 
-	public void setQty(int qty) {
+	public void setQty(float qty) {
 		this.qty = qty;
 	}
 
@@ -125,8 +123,4 @@ public class GetSellBillDetail {
 				+ cgstPer + ", itemName=" + itemName + "]";
 	}
 
-	
-	
-	
-	
 }
